+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Set `version` field of `Chart.yaml` to latest repo release fix ci checks.
+
 ## [2.2.0] - 2026-01-07
 
 ### Added